moved components into addon for possible extension. and added the ability to toggle draggability of a draggable object
