+++ addon/components/draggable-object-target.js
@@ -1,0 +1,28 @@
+import Ember from 'ember';
+import Droppable from 'ember-drag-drop/mixins/droppable';
+
+export default Ember.Component.extend(Droppable, {
+  classNames: ["draggable-object-target"],
+
+  handlePayload: function(payload) {
+    var obj = this.get('coordinator').getObject(payload,{target: this});
+    this.sendAction('action',obj,{target: this});
+  },
+
+  handleDrop: function(event) {
+    var dataTransfer = event.dataTransfer;
+    var payload = dataTransfer.getData("Text");
+    this.handlePayload(payload);
+  },
+
+  acceptDrop: function(event) {
+    this.handleDrop(event);
+  },
+
+  actions: {
+    acceptForDrop: function() {
+      var hashId = this.get('coordinator.clickedId');
+      this.handlePayload(hashId);
+    }
+  }
+});
